+++ src/core/contant.ts
@@ -1,5 +1,5 @@
 
-export const initialDefaultFrame = -999999;
+export const initialDefaultFrame = 0;
 export const degToRads = Math.PI / 180;
 
 export const defaultVector = [0, 0];
